refactor(api): Divide exception package into subpackages
