+++ backend/FleetForge/src/main/java/com/team18/FleetForge/exception/handlers/ValidationExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.exception;
+package com.team18.FleetForge.exception.handlers;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.FieldError;
