+++ backend/FleetForge/src/main/java/com/team18/FleetForge/exception/ride/RideNotFoundException.java
@@ -1,4 +1,4 @@
-package com.team18.FleetForge.exception;
+package com.team18.FleetForge.exception.ride;
 
 public class RideNotFoundException extends RuntimeException {
     public RideNotFoundException(String message) {
